backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
Werkzeug==2.3.7
Pillow==10.0.1
cloudinary==1.36.0